fix models

+++ twitoff_src/models.py
@@ -18,7 +18,8 @@ class Tweet(DB.Model):
     id = DB.Column(DB.BigInteger, primary_key=True)
     text = DB.Column(DB.Unicode(500), nullable=False)
     embedding = DB.Column(DB.PickleType, nullable=False)
-    user_id = DB.Column(DB.Integer, DB.ForeignKey('user.id'), nullable=False)
+    user_id = DB.Column(DB.BigInteger, DB.ForeignKey('user.id'),
+                        nullable=False)
     user = DB.relationship('User', backref=DB.backref('tweets', lazy=True))
 
     def __repr__(self):
